move test_unstage from test_porcelain to test_repository
