+++ dulwich/tests/test_porcelain.py
@@ -1320,89 +1320,6 @@ class ResetTests(PorcelainTestCase):
         self.assertEqual([], changes)
 
 
-class UnstageTests(PorcelainTestCase):
-    def test_unstage_midify_file_with_dir(self):
-        os.mkdir(os.path.join(self.repo.path, 'new_dir'))
-        full_path = os.path.join(self.repo.path, 'new_dir', 'foo')
-
-        with open(full_path, 'w') as f:
-            f.write('hello')
-        porcelain.add(self.repo, paths=[full_path])
-        porcelain.commit(
-            self.repo,
-            message=b"unitest",
-            committer=b"Jane <jane@example.com>",
-            author=b"John <john@example.com>",
-        )
-        with open(full_path, 'a') as f:
-            f.write('something new')
-        self.repo.unstage(['new_dir/foo'.encode()])
-        status = list(porcelain.status(self.repo))
-        self.assertEqual([{'add': [], 'delete': [], 'modify': []}, [b'new_dir/foo'], []], status)
-
-    def test_unstage_while_no_commit(self):
-        file = 'foo'
-        full_path = os.path.join(self.repo.path, file)
-        with open(full_path, 'w') as f:
-            f.write('hello')
-        porcelain.add(self.repo, paths=[full_path])
-        self.repo.unstage([file.encode()])
-        status = list(porcelain.status(self.repo))
-        self.assertEqual([{'add': [], 'delete': [], 'modify': []}, [], ['foo']], status)
-
-    def test_unstage_add_file(self):
-        file = 'foo'
-        full_path = os.path.join(self.repo.path, file)
-        porcelain.commit(
-            self.repo,
-            message=b"unitest",
-            committer=b"Jane <jane@example.com>",
-            author=b"John <john@example.com>",
-        )
-        with open(full_path, 'w') as f:
-            f.write('hello')
-        porcelain.add(self.repo, paths=[full_path])
-        self.repo.unstage([file.encode()])
-        status = list(porcelain.status(self.repo))
-        self.assertEqual([{'add': [], 'delete': [], 'modify': []}, [], ['foo']], status)
-
-    def test_unstage_modify_file(self):
-        file = 'foo'
-        full_path = os.path.join(self.repo.path, file)
-        with open(full_path, 'w') as f:
-            f.write('hello')
-        porcelain.add(self.repo, paths=[full_path])
-        porcelain.commit(
-            self.repo,
-            message=b"unitest",
-            committer=b"Jane <jane@example.com>",
-            author=b"John <john@example.com>",
-        )
-        with open(full_path, 'a') as f:
-            f.write('broken')
-        porcelain.add(self.repo, paths=[full_path])
-        self.repo.unstage([file.encode()])
-        status = list(porcelain.status(self.repo))
-        self.assertEqual([{'add': [], 'delete': [], 'modify': []}, [b'foo'], []], status)
-
-    def test_unstage_remove_file(self):
-        file = 'foo'
-        full_path = os.path.join(self.repo.path, file)
-        with open(full_path, 'w') as f:
-            f.write('hello')
-        porcelain.add(self.repo, paths=[full_path])
-        porcelain.commit(
-            self.repo,
-            message=b"unitest",
-            committer=b"Jane <jane@example.com>",
-            author=b"John <john@example.com>",
-        )
-        os.remove(full_path)
-        self.repo.unstage([file.encode()])
-        status = list(porcelain.status(self.repo))
-        self.assertEqual([{'add': [], 'delete': [], 'modify': []}, [b'foo'], []], status)
-
-
 class ResetFileTests(PorcelainTestCase):
     def test_reset_modify_file_to_commit(self):
         file = 'foo'
